Moved registries into util
