+++ src/main/java/fr/estecka/variantscit/VCitRegistries.java
@@ -20,6 +20,7 @@ import fr.estecka.variantscit.itemdata.preconditions.*;
 import fr.estecka.variantscit.modules.impl.*;
 import fr.estecka.variantscit.modules.libraries.*;
 import fr.estecka.variantscit.reload.IUnbakedModule;
+import fr.estecka.variantscit.util.DecodableRegistry;
 
 
 public final class VCitRegistries
