+++ src/main/java/fr/estecka/variantscit/trims/TrimPatternOverlay.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
-import fr.estecka.variantscit.ReloadableRepository;
+import fr.estecka.variantscit.util.ReloadableRepository;
 import net.minecraft.resources.Identifier;
 
 /**
