+++ src/main/java/fr/estecka/variantscit/util/DecodableRegistry.java
@@ -1,4 +1,4 @@
-package fr.estecka.variantscit;
+package fr.estecka.variantscit.util;
 
 import java.util.HashMap;
 import java.util.Iterator;
@@ -17,6 +17,8 @@ import com.mojang.serialization.MapDecoder;
 import com.mojang.serialization.MapEncoder;
 import com.mojang.serialization.MapLike;
 import com.mojang.serialization.RecordBuilder;
+import fr.estecka.variantscit.CodecUtil;
+import fr.estecka.variantscit.VariantsCitMod;
 
 public class DecodableRegistry<T>
 {
