+++ src/main/java/fr/estecka/variantscit/util/ReloadableRepository.java
@@ -1,10 +1,11 @@
-package fr.estecka.variantscit;
+package fr.estecka.variantscit.util;
 
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 import org.jetbrains.annotations.Nullable;
 import com.mojang.serialization.Codec;
+import fr.estecka.variantscit.CodecUtil;
 import net.minecraft.resources.Identifier;
 import net.minecraft.server.packs.resources.ResourceManager;
